pyproject.toml
[project]
name = "pdf2htmlex-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.1",
]
